Prepare_Dataset - fixed topic modelling + new pdf_to_text
There was a logic error in the method that was labelled the cases as env or non-env. I used to fix it at a later point in the code, but I have now updated the logic from the dataset creation time.

new pdf_to_text method -> many errors fixed

code cleaned and comments added